# docs/api/swagger.yml
openapi: "3.0.3"
info:
  title: Safe Travels API
  version: "1.0.0"
  description: |
    API do Safe Travels — TCC de Vicente Hofmeister (PUCRS).

    Sistema colaborativo para compartilhamento de localização em tempo real durante viagens em grupo.

    ## Status dos endpoints

    Cada endpoint indica seu estado de implementação:
    - **Sem marcação** — implementado e funcional
    - **⚠️ TODO** — planejado, ainda não implementado

    ## Autenticação

    Endpoints protegidos exigem `Authorization: Bearer <token>`.
    O token JWT é obtido via `POST /auth/register` ou `POST /auth/login`.
servers:
  - url: http://localhost:3000
    description: Desenvolvimento local (Docker Compose)
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: error
        timestamp:
          type: string
          format: date-time
        message:
          type: string
          example: "Descrição do erro"
    SuccessEnvelope:
      type: object
      properties:
        status:
          type: string
          example: ok
        timestamp:
          type: string
          format: date-time
        message:
          type: string
    PublicUser:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        username:
          type: string
          example: alice
        name:
          type: string
          example: Alice Silva
        email:
          type: string
          format: email
        createdAt:
          type: string
          format: date-time
...
    LocationData:
      type: object
      properties:
        locationEventId:
          type: integer
          example: 42
        user:
          type: object
          properties:
            userId:
              type: string
              format: uuid
            username:
              type: string
              example: alice
            name:
              type: string
              example: Alice Silva
        latitude:
          type: number
          example: -30.0574
        longitude:
          type: number
          example: -51.1778
        accuracyMeters:
          type: integer
          nullable: true
          example: 4
        capturedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
...
    GroupOwner:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        username:
          type: string
        name:
          type: string
...
    GroupData:
      type: object
      properties:
        groupId:
          type: string
          format: uuid
        name:
          type: string
          example: Viagem PUCRS 2026
        description:
          type: string
          nullable: true
        owner:
          $ref: '#/components/schemas/GroupOwner'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TripOwner:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        username:
          type: string
        name:
          type: string
...
    TripSummary:
      type: object
      properties:
        tripId:
          type: string
          format: uuid
        name:
          type: string
          example: Expedição Litoral Gaúcho 2026
        description:
          type: string
          nullable: true
        owner:
          $ref: '#/components/schemas/TripOwner'
        groupId:
          type: string
          format: uuid
          nullable: true
          description: "NULL = viagem individual; UUID = viagem de grupo"
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
          nullable: true
          description: "NULL = trip ativa"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
paths:
...
  /location/user/{userId}:
    get:
      tags: [location]
      summary: Localização mais recente de um usuário específico
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Localização encontrada
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LocationData'
        "401":
          description: Não autorizado
        "404":
          description: Nenhuma localização encontrada para este usuário

  # ─── GROUP ─────────────────────────────────────────────────────────────────
...
  /group/user/{userId}:
    get:
      tags: [group]
      summary: Lista grupos que um usuário pertence
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Lista de grupos
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          allOf:
                            - $ref: '#/components/schemas/GroupData'
                            - type: object
                              properties:
                                joinedAt:
                                  type: string
                                  format: date-time
        "401":
          description: Não autorizado
...
  /user/search:
    get:
      tags: [user]
      summary: Busca usuários por nome ou username
      parameters:
        - name: q
          in: query
          required: true
          description: Termo de busca (mínimo 2 caracteres). ILIKE sobre username e name.
          schema:
            type: string
            minLength: 2
            example: alice
      responses:
        "200":
          description: Lista de usuários encontrados (máximo 20)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/PublicUser'
        "400":
          description: Termo de busca muito curto (menos de 2 caracteres)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Não autorizado
  /user/{userId}:
    get:
      tags: [user]
      summary: Retorna perfil de um usuário
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Perfil encontrado
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PublicUser'
        "400":
          description: userId inválido
        "401":
          description: Não autorizado
        "404":
          description: Usuário não encontrado
...
  /trip/user/{userId}:
    get:
      tags: [trip]
      summary: Lista viagens de um usuário
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Lista de viagens (ativas e históricas)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/TripSummary'
        "401":
          description: Não autorizado
...